backend/core/requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
gunicorn>=21.0.0

# Authentication
djangorestframework-simplejwt>=5.2.0

# Resume Parsing - PDF/DOCX Extraction
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.0.0 
numpy>=2.2.6

# NLP - spaCy for Named Entity Recognition (Optional - Falls back to regex if not installed)
# spaCy requires Visual Studio C++ Build Tools on Windows
# Install from: https://visualstudio.microsoft.com/visual-cpp-build-tools/
# spacy>=3.7.0
# en-core-web-lg>=3.7.1

# Text Processing
regex>=2023.0.0

# Date Parsing
python-dateutil>=2.8.0

# Phone Number Parsing (Optional)
phonenumbers>=8.13.0

# Email Validation (Optional)
email-validator>=2.1.0

# Additional Utilities
pytz>=2023.3

# ML/AI - Sentence Transformers for Embeddings (bge-large-en-v1.5)
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0

# LLM Resume Parser - OpenAI client (used with GitHub Models API)
openai>=1.0.0

# LangChain for LLM-powered resume parsing
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-groq>=0.1.0

# Better PDF extraction (layout-aware, markdown output)
pymupdf4llm>=0.0.17
